chore: file restructuring
