+++ app/src/features/apiClient/errors/FsError/FsAccessError/components/FsAccessTroubleshoot/FsAccessTroubleshoot.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Snippet } from "componentsV2/Snippet/Snippet";
-import { FsAccessError } from "features/apiClient/errors/FsError/FsAccessError";
+import { FsAccessError } from "features/apiClient/errors/FsError/FsAccessError/FsAccessError";
 import "./fsAccessTroubleshoot.scss";
 
 interface Props {
@@ -15,13 +15,16 @@ export const FsAccessTroubleshoot: React.FC<Props> = ({ error }) => {
       </div>
       <Snippet code={`chown $(whoami) ${error.path}`} allowCopy />
       <div className="fs-access-troubleshoot__text" style={{ margin: "16px 0" }}>
-        To learn more about{" "}
-        <span className="error-path">chown</span> refer{" "}
-        <a href="https://pubs.opengroup.org/onlinepubs/9799919799/utilities/chown.html" target="_blank" rel="noreferrer">
+        To learn more about <span className="error-path">chown</span> refer{" "}
+        <a
+          href="https://pubs.opengroup.org/onlinepubs/9799919799/utilities/chown.html"
+          target="_blank"
+          rel="noreferrer"
+        >
           here
         </a>
       </div>
-      <br/>
+      <br />
       <div className="fs-access-troubleshoot__text" style={{ margin: "16px 0" }}>
         If the above command doesn't work, you can try creating a new folder of which you will have write access and
         copying the files to it :
