+++ app/src/features/apiClient/errors/FsError/FsAccessError/components/FsAccessTroubleshoot/fsAccessTroubleshoot.scss
@@ -1,7 +1,7 @@
 .fs-access-troubleshoot {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
   &__text {
     font-size: var(--requestly-font-size-sm);
     color: var(--requestly-color-text-subtle);
